fixed int width

+++ bms/App/lib/gui/inc/gui_drivers.h
@@ -15,10 +15,10 @@ void process_can_command(uint32_t ext_id, uint8_t *data);
 
 void send_can_error(can_error_id_t error_id);
 
-void cell_voltage_readings(cell_asic_ctx_t *asic, int start_ic, int end_ic,
+void cell_voltage_readings(cell_asic_ctx_t *asic, uint8_t start_ic, uint8_t end_ic,
                            uint8_t *data_arr);
 
-void therm_temp_readings(cell_asic_ctx_t *asic, int start_ic, int end_ic,
+void therm_temp_readings(cell_asic_ctx_t *asic, uint8_t start_ic, uint8_t end_ic,
                          uint8_t *data_arr);
 
 void metadata_readings(pack_data_t *pack, pcb_ctx_t *pcb, uint8_t *data_arr);
